+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+## 0.0.11
+
+Features
+
+* Added `#sanitize` which returns the formatted identifier or `nil` if invalid
+
+Bugfixes
+
+* Fixed a rare edge case bug that would not properly format a Swedish Identifier such as "0asdfghj501261853"
+
 ## 0.0.10
 
 *   No more Ruby 1.8.7 support
-*   Introduced `#age` for national identificaiton numbers
+*   Introduced `#age` for national identification numbers
 
 ## 0.0.9